requirements.txt
langgraph-checkpoint-redis>=0.1.0
langgraph-store-redis>=0.1.0
redis>=5.0.0
pyyaml>=6.0
typer>=0.12.0
chromadb>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
langchain-google-genai>=1.0.0
apscheduler>=3.10.0
pytz>=2024.1
python-telegram-bot>=20.0
google-genai>=0.1.0
langgraph>=0.0.10
langchain-core>=0.1.0
pydantic>=2.0.0
